requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
PyPDF2
python-docx
scikit-learn
joblib
pandas
google-genai
# Auth
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
google-auth
# Add other dependencies here
